Big renaming and classes' organizing
