+++ qwerty/Form1.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Threading;
 using System.Media;
+using qwerty.Objects;
 
 namespace qwerty
 {
