+++ qwerty/Objects/Meteor.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Drawing;
+﻿using System.Drawing;
 
-namespace qwerty
+namespace qwerty.Objects
 {
-    class Meteor : SpaceShit
+    class Meteor : SpaceObject
     {
         public int explodeDmg;
         public string staticDescription = "Движущийся метеор";
