+++ qwerty/Objects/ShipAssaulter.cs
@@ -1,11 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Drawing;
+using qwerty.Objects.Weapons;
 
-namespace qwerty
+namespace qwerty.Objects
 {
     class ShipAssaulter : Ship
     {
