+++ qwerty/Objects/Weapons/LightLaser.cs
@@ -1,16 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Drawing;
+﻿using System.Drawing;
 using System.Windows.Forms;
 
-namespace qwerty
+namespace qwerty.Objects.Weapons
 {
-    class wpnLightLaser : Weapon
+    class LightLaser : Weapon
     {
-        public wpnLightLaser()
+        public LightLaser()
         {
             attackPower = 25;
             attackRange = 3;
