+++ qwerty/Objects/Weapons/Weapon.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
+﻿using System.Windows.Forms;
 
-namespace qwerty
+namespace qwerty.Objects.Weapons
 {
     public enum WeaponType
     {
