+++ qwerty/qwerty.csproj
@@ -53,18 +53,18 @@
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
-    <Compile Include="Meteor.cs" />
+    <Compile Include="Objects\Meteor.cs" />
     <Compile Include="ObjectManager.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Ship.cs" />
-    <Compile Include="ShipAssaulter.cs" />
-    <Compile Include="ShipScout.cs" />
-    <Compile Include="SpaceShit.cs" />
-    <Compile Include="Weapon.cs" />
-    <Compile Include="WpnHeavyLaser.cs" />
-    <Compile Include="WpnLightIon.cs" />
-    <Compile Include="wpnLightLaser.cs" />
+    <Compile Include="Objects\Ship.cs" />
+    <Compile Include="Objects\ShipAssaulter.cs" />
+    <Compile Include="Objects\ShipScout.cs" />
+    <Compile Include="Objects\SpaceObject.cs" />
+    <Compile Include="Objects\Weapons\Weapon.cs" />
+    <Compile Include="Objects\Weapons\HeavyLaser.cs" />
+    <Compile Include="Objects\Weapons\LightIon.cs" />
+    <Compile Include="Objects\Weapons\LightLaser.cs" />
     <EmbeddedResource Include="Form1.resx">
       <DependentUpon>Form1.cs</DependentUpon>
     </EmbeddedResource>
@@ -90,6 +90,7 @@
   <ItemGroup>
     <None Include="App.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
